Refactor UI structure and consolidate webview components
- Moved webview-related files to a new `src/ui/webview` directory for better organization.
- Created a new `SidebarProvider` class to manage the sidebar webview.
- Implemented a shared types file for consistent messaging between webviews and the extension.
- Removed old webview types and manager files, replacing them with updated implementations.
- Updated Vite configuration to reflect new file paths for webview assets.
- Enhanced sidebar and webview HTML, CSS, and TypeScript for improved functionality and styling.
- Added tests for the new `WebviewManager` class to ensure proper initialization and disposal.
