+++ src/extension.ts
@@ -1,8 +1,8 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
-import { WebviewManager } from './webview/WebviewManager';
-import { SidebarProvider } from './sidebar/SidebarProvider';
+import { WebviewManager } from './ui/webview/WebviewManager';
+import { SidebarProvider } from './ui/sidebar/SidebarProvider';
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
